+++ main-server/src/core/ndid/index.js
@@ -60,10 +60,10 @@ export async function initNDID({
     await tendermint.transact({
       nodeId: config.nodeId,
       fnName: 'InitNDID',
-      params: { 
-        node_id: 'ndid1', 
-        public_key, 
-        master_public_key, 
+      params: {
+        node_id: config.nodeId,
+        public_key,
+        master_public_key,
         chain_history_info: JSON.stringify(chain_history_info),
       },
     });