+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## TBD
 
+BREAKING CHANGES:
+
+- Remove `SKIP_GET_ROLE` environment variable option. Add `NDID_NODE` environment variable option for skipping get role and wait for blockchain initialization to end.
+
 BUG FIXES:
 
 - Fix load and retry backlog transact requests is called when server initialization failed.
 - Wait for blockchain initialization to end on server start.
 - Fix request timeout is called before blockchain initialization ended while migrating to a new chain.
+- Fix NDID node cannot be started if blockchain initialization is not ended.
 
 ## 0.11.4 (November 12, 2018)